@@ -1,32 +1,31 @@
-namespace MathExpressionParser.Console
+namespace MathExpressionParser.Console;
+
+using System;
+using MathExpressionParser;
+
+public class Program
 {
-    using System;
-    using MathExpressionParser;
-
-    public class Program
+    public static void Main()
     {
-        public static void Main()
+        Console.WriteLine("Please, type in an arithmetic expression:");
+
+        try
         {
-            Console.WriteLine("Please, type in an arithmetic expression:");
+            string? input = Console.ReadLine();
 
-            try
+            if (string.IsNullOrEmpty(input))
             {
-                var input = Console.ReadLine();
-
-                if (string.IsNullOrEmpty(input))
-                {
-                    Console.WriteLine("The input string is empty!");
-                    return;
-                }
-
-                decimal result = MathProcessor.Calculate(input);
-                Console.WriteLine($"Expression result: {result}");
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine(ex.Message);
+                Console.WriteLine("The input string is empty!");
                 return;
             }
+
+            decimal result = MathProcessor.Calculate(input);
+            Console.WriteLine($"Expression result: {result}");
+        }
+        catch (Exception ex)
+        {
+            Console.WriteLine(ex.Message);
+            return;
         }
     }
 }
